move Twinmaker creation in nested stack
